fix(opencode): satisfy source size guard
